// jcpm-frontend/src/schemas/registerSchema.ts
import * as yup from 'yup';

export const registerSchema = yup.object().shape({
  username: yup.string().required("Username is required"),
  email: yup.string().email("Invalid email").required("Email is required"),
  password: yup.string().min(6, "Password must be at least 6 characters").required("Password is required"),
  password2: yup.string()
    .oneOf([yup.ref('password')], "Passwords do not match")
    .required("Please confirm your password"),
});